Comments fix
- [x] Removed the constructor. Title is now constant.
- [x] Created class HomeScreen, and moved HomePage there.
- [x] Renamed the variable.
